Clean_up_and_good_file
