+++ servo_motor.py
@@ -1,13 +1,13 @@
 import pigpio
 from time import sleep
 
-servo = 17 # Le numéro GPIO de la broche ou est connecté le servomoteur
-FREQ = 50 # Fréquence en Hz de la période
+servo = 17
+FREQ = 50 
 
 pi = pigpio.pi()
 pi.set_mode(servo,pigpio.OUTPUT)
 pi.set_PWM_frequency(servo,FREQ)
-pi.set_PWM_range(servo,100) # Valeurs possibles dutycycle de 0-100
+pi.set_PWM_range(servo,100)
 
 plus90 = 12.5
 moins90 = 2.5
